+++ src/main/java/com/gae4coon/cloudmaestro/domain/ssohost/service/SecurityGroupService.java
@@ -1,16 +1,131 @@
 package com.gae4coon.cloudmaestro.domain.ssohost.service;
 
+import com.gae4coon.cloudmaestro.domain.resource.service.AddResourceService;
 import com.gae4coon.cloudmaestro.domain.ssohost.dto.GroupData;
 import com.gae4coon.cloudmaestro.domain.ssohost.dto.NodeData;
 import com.gae4coon.cloudmaestro.domain.ssohost.dto.LinkData;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+@Service
+@RequiredArgsConstructor
+public class SecurityGroupService{
+    private final DiagramDTOService diagramDTOService;
+    private final AddResourceService addResourceService;
 
-public interface SecurityGroupService {
-    void addSecurityGroup(List<NodeData> nodeDataList, List<GroupData> groupDataList, List<LinkData> linkDataList);
-    void modifySecurityGroupLink(List<NodeData> nodeDataList, List<GroupData> groupDataList, List<LinkData> linkDataList);
+    public void addSecurityGroup(List<NodeData> nodeDataList, List<GroupData> groupDataList, List<LinkData> linkDataList) {
+        List<NodeData> fwNodeData = diagramDTOService.getNodeListByText(nodeDataList, "Firewall");
+
+        for(NodeData fw: fwNodeData) {
+            String fwVPC = null;
+            List<LinkData> fwLinkData = diagramDTOService.getLinkDataListByFrom(linkDataList, fw.getKey());
+            System.out.println("fwLinkData" + fwLinkData);
+            GroupData sg = addResourceService.addSecurityGroup(groupDataList);
+
+            for(LinkData fwLink: fwLinkData) {
+                System.out.println("fwLink" + fwLink);
+
+                // FW -> Node
+                NodeData toNode = diagramDTOService.getNodeDataByKey(nodeDataList, fwLink.getTo());
+                // server, database인 경우 Security Group 처리
+                System.out.println("toNode" + toNode);
+                if(toNode!=null && (toNode.getText().contains("Server") || toNode.getText().contains("Database"))) {
+                    if(!diagramDTOService.isGroupDataEquals(groupDataList, sg.getKey())) {
+                        sg.setGroup(toNode.getGroup());
+                        groupDataList.add(sg);
+                    }
+                    toNode.setGroup(sg.getKey());
+                }
+
+                // FW -> Group
+                GroupData toGroup = diagramDTOService.getGroupDataByKey(groupDataList, fwLink.getTo());
+                // Network Firewall 처리
+                if(toGroup!=null) {
+                    System.out.println("toGroup" + toGroup);
+                    if(fwVPC!=null){
+                        toGroup.setGroup(fwVPC);
+                        continue;
+                    };
+
+                    GroupData vpc = addResourceService.addVPC(groupDataList);
+                    vpc.setGroup("Region");
+                    groupDataList.add(vpc);
+                    fwVPC = vpc.getKey();
+
+                    toGroup.setGroup(vpc.getKey());
+
+                    NodeData NetworkFW = diagramDTOService.getNodeDataByText(nodeDataList, "Network Firewall");
+                    // NF 없으면 생성
+                    if (NetworkFW == null) {
+                        // add Network Firewall into Region
+                        NetworkFW = addResourceService.addNetworkFirewall();
+                        NetworkFW.setGroup("Region");
+                        NetworkFW.setLoc("0 -300");
+                        nodeDataList.add(NetworkFW);
+                    }
+
+                    // create Network Firewall Group in VPC
+                    GroupData fps = addResourceService.addPublicsubnet();
+                    fps.setKey("Firewall Public Subnet"+diagramDTOService.getGroupNumber(groupDataList, "Firewall Public Subnet"));
+                    fps.setText("Firewall Public Subnet");
+                    fps.setGroup(vpc.getKey());
+                    groupDataList.add(fps);
+
+                    // add Network Firewall endpoint into Firewall Public Subnet
+                    NodeData fwend = addResourceService.addNetworkFirewallEndpoints(nodeDataList);
+                    fwend.setGroup(fps.getKey());
+                    fwend.setLoc("200 -200");
+                    nodeDataList.add(fwend);
+
+                    LinkData netFWtoendpoint = new LinkData();
+                    netFWtoendpoint.setFrom(NetworkFW.getKey());
+                    netFWtoendpoint.setTo(fwend.getKey());
+                    linkDataList.add(netFWtoendpoint);
+
+                    System.out.println("netFWtoendpoint" + netFWtoendpoint);
+
+                }
+            }
+        }
+    }
+
+    public void modifySecurityGroupLink(List<NodeData> nodeDataList, List<GroupData> groupDataList, List<LinkData> linkDataList) {
+        for (NodeData nodeData : nodeDataList) {
+            if (nodeData.getGroup()!=null&&!nodeData.getGroup().contains("Security Group")) continue;
+
+            for (LinkData linkData : linkDataList) {
+                if (linkData.getFrom()!=null && linkData.getFrom().equals(nodeData.getKey())) {
+                    linkData.setFrom(nodeData.getGroup());
+                } else if (linkData.getTo()!=null && linkData.getTo().equals(nodeData.getKey())) {
+                    linkData.setTo((nodeData.getGroup()));
+
+                }
+            }
+        }
+
+        Map<List<GroupData>, List<NodeData>> resultMap = new HashMap<>();
+        resultMap.put(groupDataList, nodeDataList);
+    }
 
 
-    void addEc2Group(List<NodeData> nodeDataList, List<GroupData> groupDataList, List<LinkData> linkDataList);
+    // from 이 server이고, to가 같은 목적지라면 ?
+    public void addEc2Group(List<NodeData> nodeDataList, List<GroupData> groupDataList, List<LinkData> linkDataList){
+
+        List<NodeData> newNodeDataList = new ArrayList<>();
+        List<GroupData> newGroupDataList = new ArrayList<>();
+
+        // Server에서 같은 to를 향한다면 같은 그룹으로 묶기
+        for (NodeData nodedata : nodeDataList) {
+            System.out.println("NodeData" + nodedata);
+            for(LinkData linkdata : linkDataList){
+                if(nodedata.getText().contains("Server")){
+
+
+                }
+            }
+        }
+
+
+    }
 }
