+++ src/main/resources/application.properties
@@ -4,6 +4,9 @@ spring.jwt.refreshKey=acd008377d08153f6fe2a1866a73e4b69d8e7aa86e1d1e66b7afbff593
 spring.jwt.token.access-expiration-time=43200000
 spring.jwt.token.refresh-expiration-time=604800000
 
+logging.file.name=logs/myapp.log
+
+
 # AWS S3 - cm-user-file
 cloud.aws.s3.bucket=cm-user-file
 cloud.aws.credentials.accessKey=AKIAQ4MLWFLEVAWTZ4DS
